requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
lxml==6.0.0
nltk==3.8.1
numpy==2.3.2
openai==1.97.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
realtime==2.6.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==15.0.1
xlsxwriter==3.2.5
tiktoken==0.7.0
